Bean Validation

+++ src/main/java/com/hnrqne/crudclient/services/ClientService.java
@@ -1,6 +1,7 @@
 package com.hnrqne.crudclient.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -9,7 +10,10 @@ import org.springframework.transaction.annotation.Transactional;
 import com.hnrqne.crudclient.dto.ClientDTO;
 import com.hnrqne.crudclient.entities.Client;
 import com.hnrqne.crudclient.repositories.ClientRepository;
+import com.hnrqne.crudclient.services.exceptions.DatabaseException;
 import com.hnrqne.crudclient.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ClientService {
@@ -43,16 +47,29 @@ public class ClientService {
 	
 	@Transactional
 	public ClientDTO update(Long id, ClientDTO dto) {
-		Client entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
 		
-		entity = repository.save(entity);
-		
-		return new ClientDTO(entity);
+		try {
+			Client entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			
+			entity = repository.save(entity);
+			
+			return new ClientDTO(entity);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		} 
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
 	}
 	
 	private void copyDtoToEntity(ClientDTO dto, Client entity) {
